requirements.txt
decorator==4.0.10
ecdsa==0.13
Fabric3==1.12.post1
ipdb==0.10.1
ipython==5.1.0
ipython-genutils==0.1.0
paramiko==1.18.1
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.9
ptyprocess==0.5.1
pycrypto==2.6.1
Pygments==2.1.3
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.1
wcwidth==0.1.7
